+++ cmd/template/cluster/provider/capa.go
@@ -11,6 +11,8 @@ import (
 	"github.com/giantswarm/microerror"
 	"sigs.k8s.io/yaml"
 
+	k8smetadata "github.com/giantswarm/k8smetadata/pkg/label"
+
 	"github.com/giantswarm/kubectl-gs/cmd/template/cluster/provider/templates/aws"
 	"github.com/giantswarm/kubectl-gs/cmd/template/cluster/provider/templates/capa"
 	"github.com/giantswarm/kubectl-gs/internal/key"
@@ -130,6 +132,8 @@ func templateClusterAWS(ctx context.Context, k8sClient k8sclient.Interface, outp
 			return microerror.Mask(err)
 		}
 
+		userConfigMap.ObjectMeta.Labels[k8smetadata.Cluster] = config.Name
+
 		configMapYAML, err = yaml.Marshal(userConfigMap)
 		if err != nil {
 			return microerror.Mask(err)
@@ -155,6 +159,9 @@ func templateClusterAWS(ctx context.Context, k8sClient k8sclient.Interface, outp
 			Namespace:               organizationNamespace(config.Organization),
 			Version:                 appVersion,
 			UserConfigConfigMapName: configMapName,
+			ExtraLabels: map[string]string{
+				k8smetadata.Cluster: config.Name,
+			},
 		}
 
 		var err error
@@ -198,6 +205,8 @@ func templateDefaultAppsAWS(ctx context.Context, k8sClient k8sclient.Interface, 
 			return microerror.Mask(err)
 		}
 
+		userConfigMap.ObjectMeta.Labels[k8smetadata.Cluster] = config.Name
+
 		configMapYAML, err = yaml.Marshal(userConfigMap)
 		if err != nil {
 			return microerror.Mask(err)
@@ -224,6 +233,9 @@ func templateDefaultAppsAWS(ctx context.Context, k8sClient k8sclient.Interface, 
 			Namespace:               organizationNamespace(config.Organization),
 			Version:                 appVersion,
 			UserConfigConfigMapName: configMapName,
+			ExtraLabels: map[string]string{
+				k8smetadata.Cluster: config.Name,
+			},
 		})
 		if err != nil {
 			return microerror.Mask(err)